FuncType

Documentation for FuncType.

FuncType.@Func

FuncType.@functype

FuncType.@Func — Macro

@Func(<func type sig>)

Create a FunctionType with function type signature syntax. Because Julia function are not curried, function with multiple argument need to be explicitly wrap with parenthesis. On the other hand, Tuple input need to be specified as something like Tuple{A, B}.

Example

julia> @Func Int -> Int   # single input single output
FunctionType{Int64, Tuple{Int64}}

julia> @Func (Int, Float64) -> Float64   # multiple input single output
FunctionType{Float64, Tuple{Int64, Float64}}

julia> @Func Int -> (Float64, Float64)   # single input multiple output
FunctionType{Tuple{Float64, Float64}, Tuple{Int64}}

julia> @Func (Int, Int) -> (Float64, Float64)   # multiple input multiple output
FunctionType{Tuple{Float64, Float64}, Tuple{Int64, Int64}}

julia> @Func () -> Int   # no input
FunctionType{Int64, Tuple{}}

julia> @Func Tuple{A, A} -> A   # single tuple input
FunctionType{var"#41#A", Tuple{Tuple{var"#41#A", var"#41#A"}}} where var"#41#A"

julia> @Func (Tuple{A, B}, B) -> C   # multiple input with tuple argument
FunctionType{<:Any, Tuple{Tuple{var"#42#A", var"#43#B"}, var"#43#B"}} where {var"#42#A", var"#43#B"}

julia> @Func T -> T   # type var
FunctionType{var"#30#T", Tuple{var"#30#T"}} where var"#30#T"

julia> @Func @NamedTuple{x::A, y::A} -> A   # nested macro
FunctionType{var"#40#A", Tuple{NamedTuple{(:x, :y), Tuple{var"#40#A", var"#40#A"}}}} where var"#40#A"

julia> @Func (@Func(A -> B), Vector{A}) -> Vector{B}   # higher-order function: `map`
FunctionType{Vector{var"#15#B"}, Tuple{FunctionType{var"#15#B", Tuple{var"#14#A"}}, Vector{var"#14#A"}}} where {var"#14#A", var"#15#B"}

julia> @Func ((C, A)->(C, B), C, Vector{A}) -> (C, Vector{B})   # scan :: (c -> a -> (c, b)) -> c -> [a] -> (c, [b])
FunctionType{Tuple{var"#47#C", Vector{var"#46#B"}}, Tuple{FunctionType{Tuple{var"#47#C", var"#46#B"}, Tuple{var"#47#C", var"#45#A"}}, var"#47#C", Vector{var"#45#A"}}} where {var"#45#A", var"#46#B", var"#47#C"}

FuncType.@functype — Macro

@functype function <func name>(<arg with type sig>)::<ret type> where {<parameters, if any>}
    <func body>
end

If ret type is missing, assuming the function return nothing